Fix for use-case where LOG_LEVEL is not overridden
I see the appeal in ditching the entrypoint script since it only existed to allow running things without `LOG_LEVEL` overridden, but you'll need these changes in here or else the user _must_ defined `LOG_LEVEL` on their machine to run things. One thing that can be confusing is where ENV vars come from in the context of docker-compose.

Although your line under `shared_environment` successfuly sets `LOG_LEVEL` _within_ those services, the references to `$LOG_LEVEL` in the run commands are resolved by Docker when reading this compose file and refer to the host environment, not the service's soon-to-be environment.