Adjust for local dev env
In a dev environment it is _likely_ preferable not to require the user to bind mount their own SSL cert into the mysql service. One of the big differences between the `mysql/mysql-server` and `mysql` docker images is that the latter has the ability to generate a self-signed cert when it starts up.

Of course, for Vapor to communicate with it, the mysql config still needs to be modified to not use a certificate signing authority.